requirements.txt
fastapi
uvicorn[standard]
python-dotenv
crewai
langchain-mistralai
pandas
python-multipart
aiofiles
showdown
openpyxl
gunicorn